BACKEND/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.11.5",
    "chromadb>=1.4.1",
    "dnspython>=2.7.0",
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "motor>=3.7.1",
    "playwright>=1.58.0",
    "pymongo>=4.16.0",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.2",
    "python-multipart>=0.0.22",
    "python-docx>=1.1.2",
    "python-pptx>=1.0.2",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.40.0",
]
